refactor: 불필요한 리소스 및 폴더 정리
