--- src/app/(auth)/kakao/redirect/page.module.scss
@@ -1,7 +1,0 @@
-.spinnerWrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  width: 100%;
-}
