--- src/app/(auth)/signin/page.module.scss
@@ -1,21 +1,0 @@
-.layout {
-  height: calc(100vh - 70px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.login_container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  gap: 20px;
-  border-radius: 12px;
-  box-shadow: 0 0 4px 2px rgba(1, 1, 1, 0.2);
-}
-
-.title {
-  font-size: 25px;
-  font-weight: 600;
-}
