--- src/components/auth/LoginModal/LoginModal.module.scss
@@ -1,35 +1,0 @@
-.modal_layout {
-  padding: 20px;
-  background-color: #f0f4f8;
-}
-
-.title {
-  text-align: center;
-  font-size: 25px;
-  font-weight: 600;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  padding: 20px 40px 40px !important;
-
-  .kakao_wrap {
-    margin-top: 24px;
-    width: 100%;
-    height: 50px;
-    position: relative;
-  }
-
-  .kakao_wrap_not_active {
-    cursor: default;
-    opacity: 0.5;
-    margin-top: 24px;
-    width: 100%;
-    height: 50px;
-    position: relative;
-  }
-}
